third_party/WebKit/LayoutTests/platform/chromium-mac-snowleopard/editing/pasteboard/drop-file-svg-expected.txt

CONSOLE MESSAGE: line 1: Uncaught ReferenceError: start is not defined
CONSOLE MESSAGE: line 2: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 3: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 4: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 5: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 6: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 7: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 8: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 9: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 10: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 11: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 12: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 13: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 14: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 15: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 16: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 17: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 18: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 19: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 20: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 21: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 22: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 23: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 24: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 25: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 26: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 27: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 28: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 29: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 30: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 31: Uncaught ReferenceError: addRow is not defined
CONSOLE MESSAGE: line 32: Uncaught ReferenceError: addRow is not defined
